+++ src/js/others.js
@@ -40,17 +40,22 @@ export default {
     this.clearEnforceFocus();
     addListener(document, EVENT_FOCUSIN, (this.onFocusin = (event) => {
       const { viewer } = this;
-      const { target } = event;
+      let { target } = event;
 
-      if (target !== document
-        && target !== viewer
-        && !viewer.contains(target)
-
-        // Avoid conflicts with other modals (#474)
-        && (target.getAttribute('tabindex') === null || target.getAttribute('aria-modal') !== 'true')
-      ) {
-        viewer.focus();
+      if (target === document || target === viewer || viewer.contains(target)) {
+        return;
       }
+
+      while (target) {
+        // Avoid conflicts with other modals (#474, #540)
+        if ((target.getAttribute('tabindex') !== null || target.getAttribute('aria-modal') === 'true')) {
+          return;
+        }
+
+        target = target.parentElement;
+      }
+
+      viewer.focus();
     }));
   },